Refactor Policy

+++ app/Core/User/UserCore.php
@@ -7,8 +7,10 @@ namespace App\Core\User;
 use App\Core\Formatter\ExceptionMessage\ExceptionMessageStandard;
 use App\Core\Query\Enum\OrderDirection;
 use App\Core\User\Query\UserOrderBy;
+use App\Exceptions\Core\Auth\Permission\InsufficientPermissionException;
 use App\Exceptions\Core\User\UserEmailDuplicatedException;
-use App\Models\User\Enum\UserExceptionCode;
+use App\Models\Permission\Enum\RoleName;
+use App\Core\User\Enum\UserExceptionCode;
 use App\Models\User\User;
 use App\Port\Core\User\CreateUserPort;
 use App\Port\Core\User\DeleteUserPort;
@@ -44,6 +46,8 @@ class UserCore implements UserCoreContract
             $user->password = Hash::make($request->getUserPassword());
             $user->save();
 
+            $user->syncRoles(RoleName::NORMAL);
+
             DB::commit();
 
             return $user;
@@ -56,16 +60,39 @@ class UserCore implements UserCoreContract
     public function delete(DeleteUserPort $request): void
     {
         $user = $request->getUserModel();
+
+        if ($request->getUserActor()->cannot('delete', $user)) {
+            throw new InsufficientPermissionException(new ExceptionMessageStandard(
+                'Insufficient permission to delete user',
+                UserExceptionCode::PERMISSION_INSUFFICIENT->value,
+            ));
+        }
+
         $user->delete();
     }
 
     public function get(GetUserPort $request): User
     {
-        return $request->getUserModel();
+        $user = $request->getUserModel();
+        if ($request->getUserActor()->cannot('see', $user)) {
+            throw new InsufficientPermissionException(new ExceptionMessageStandard(
+                'Insufficient permission to see user',
+                UserExceptionCode::PERMISSION_INSUFFICIENT->value,
+            ));
+        }
+
+        return $user;
     }
 
     public function getAll(GetAllUserPort $request): LengthAwarePaginator
     {
+        if ($request->getUserActor()->cannot('seeAll', User::class)) {
+            throw new InsufficientPermissionException(new ExceptionMessageStandard(
+                'Insufficient permission to see all users',
+                UserExceptionCode::PERMISSION_INSUFFICIENT->value,
+            ));
+        }
+
         $page = $request->getPage() ?? 1;
         $perPage = $request->getPerPage() ?? 30;
         $orderDirection = $request->getOrderDirection() ?? OrderDirection::DESCENDING;
@@ -82,8 +109,14 @@ class UserCore implements UserCoreContract
             DB::beginTransaction();
 
             $user = $request->getUserModel();
-            $email = $request->getEmail();
+            if ($request->getUserActor()->cannot('update', $user)) {
+                throw new InsufficientPermissionException(new ExceptionMessageStandard(
+                    'Insufficient permission to update user',
+                    UserExceptionCode::DUPLICATED->value,
+                ));
+            }
 
+            $email = $request->getEmail();
             $isEmailExist = User::query()
                 ->where('email', $email)
                 ->where('id', '<>', $user->id)
